+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 1.0.0 - 2016-10-07
+
+### Changed
+
+- Renamed the `Memory` property to `MaxMemory`
+
+### Added
+
+- The memory reservation can now be specified using the `ReservedMemory` property -
+
+    ```xml
+    <profile pluginId="cd.go.contrib.elastic-agent.docker" id="foo">
+      <property>
+        <key>ReservedMemory</key>
+        <!-- You may use prefix B, KB, MB, GB -->
+        <value>1024MB</value>
+      </property>
+    </profile>
+    ```
+
+
 ## 0.3.1 - 2016-10-07
 
 ### Fixed